Switch data input to a large matrix rather than indiviual arrays

+++ fel_input_struc.h
@@ -1,11 +1,11 @@
 struct intergrator_input {
 	int N_theta;	// Points in theta
 	int N_p;	// Electron per theta soace	
-	float off_p;	// P offset
-	float sigma;	// Energy spread
-	float z_0;	// z start
-	float z_f;	// z end
+	double off_p;	// P offset
+	double sigma;	// Energy spread
+	double z_0;	// z start
+	double z_f;	// z end
 	int z_num;
-	float a_0;
-	float phi_0;
+	double a_0;
+	double phi_0;
 };
